add sever.js
