+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "2ndportfolio",
+  "version": "0.9.0",
+  "description": "Jeremiah Walters' portfolio which features his skillset, projects and blog entries",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jtwalters25/2ndportfolio.git"
+  },
+  "keywords": [
+    "Jeremiah",
+    "Portfolio"
+  ],
+  "author": "Jeremiah Walters",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/jtwalters25/2ndportfolio/issues"
+  },
+  "homepage": "https://github.com/jtwalters25/2ndportfolio#readme"
+}
