+++ public/index.html
@@ -175,7 +175,7 @@
   <script src="/vendor/scripts/highlight.pack.js"></script>
 
   <script src="scripts/blogArticles.js" type="text/javascript"></script>
-  <!-- <script src="scripts/projectData.js" type="text/Javascript"></script> -->
+
   <script src="scripts/projects.js" type="text/Javascript"></script>
   <script src="scripts/blogs.js" type="text/Javascript"></script>
   <script src="scripts/mainView.js" type="text/Javascript"></script>
