+++ public/scripts/blogArticles.js
@@ -1,0 +1,52 @@
+'use strict';
+var blogArticles = [
+  {'title':'Migos','body':'<p>Bad&quot; is the first Hot 100 No. 1 for both hip-hop trio Migos (Quavo, Takeoff and Offset) and rapper Lil Uzi Vert (born Symere Woods). It&#39;s also the first Hot 100 leader for the 300 Entertainment label, founded in 2014, as well as the first topper for the Quality Control imprint. (In the Hot 100&#39;s 58-year history, the song is the 1,060th No. 1 all-time.)</p>\n<p>As previously reported, &quot;Bad&quot; has inspired numerous memes and tweets playing off the song&#39;s &quot;raindrop, drop top&quot; lyrics. A 60-second clip of Migos of performing &quot;Bad&quot; in Lagos, Nigeria, posted to Twitter Dec. 20, has also helped the song&#39;s profile surge.</p>\n<p>&quot;Bad&quot; tops the Streaming Songs chart for a second week with 35.3 million U.S. streams (up 9 percent) in the week ending Jan. 5, according to Nielsen Music. The majority of its streams stem from YouTube (19.3 million), followed by Spotify (9.2 million). The track also leads the audio subscription services-based On-Demand Streaming Songs chart for a second week (15.9 million on-demand clicks, up 7 percent).</p>\n<p>&quot;Bad&quot; pushes 4-2 in Digital Song Sales (76,000 downloads sold, down 35 percent; most songs dropped significantly in sales in the tracking week, following a hearty prior sales frame that included Christmas-week shopping) and nears the Radio Songs chart, gaining by 34 percent to 20 million in airplay audience. On R&amp;B/Hip-Hop Airplay, it bounds 14-6 as the chart&#39;s Greatest Gainer (up 31 percent to 16.2 million in audience at the format).</p>\n<p>&quot;Bad&quot; additionally takes over atop Billboard&#39;s Hot R&amp;B/Hip-Hop Songs and Hot Rap Songs charts.</p>\n<p>Looking ahead to next week, &quot;Bad&quot; could further gain thanks to Donald Glover&#39;s shout-out of the song during his acceptance speech at the Golden Globes last night (Jan. 8). With Migos having guested on the FX series Atlanta, created by and starring Glover, he praised, &quot;I really wanna thank the Migos, not for being in the show, but for making &#39;Bad and Boujee.&#39; Like, that&#39;s the best song ever.&quot; (Any consumer reaction to the mention would count for next week&#39;s Hot 100, which will cover streaming and sales from Jan. 6 through Jan. 12.)</p>\n<p>Offset, Quavo, and Takeoff of Migos pose backstage during the kick off the 2016 Honda Civic Tour: Future Now at Philips Arena on June 29, 2016 in Atlanta.<br>READ MORE\nMigos&#39; &#39;Bad And Boujee&#39; Is the Latest Meme Soundtrack</p>\n','projectUrl':'http://www.billboard.com/articles/columns/chart-beat/7648006/hot-100-migos-bad-and-boujee-lil-uzi-vert-number-one','publishedOn':'2017-01-10T18:00:45.995Z','daysAgo':0,'publishedStatus':'published 0 days ago'},
+  {
+    title: 'My Journey',
+    blogcategory: 'Finances',
+    publishedOn : '10/10/2016',
+    body: '<p>Hello World!  My name is Jeremiah Walters I am a proud father and committed student in need of financial support to complete my education as a coding and computer programming student.</p>'
+  },
+  {
+    title: 'J. Cole',
+    blogcategory: 'Music & Entertainment',
+    publishedOn : '7/10/2016',
+    body: '<p>Jermaine Lamarr Cole (born January 28, 1985), known professionally as J. Cole, is an American rapper, songwriter and record producer.</p>'
+  },
+  {
+    title: 'Seattle Start Ups',
+    blogcategory: 'Tech',
+    publishedOn : '6/10/2016',
+    body: '<p>Seattle startup culture is booming and its good for people who are interested in running a business that allows them to be creative with their approach.</p>'
+  },
+  {
+    title: '1st 5k',
+    blogcategory: 'Health & Exercise',
+    publishedOn : '5/10/2016',
+    body: '<p>Running is good for your heart:)</p>'
+  },
+  {
+    title: 'No More Chicken Nuggets',
+    blogcategory: 'Health & Exercise',
+    publishedOn : '4/10/2016',
+    body: '<p>eat better live longer</p>'
+  },
+  {
+    title: 'Traffic Sucks! Can I Just Stay Home?',
+    blogcategory: 'Lifestyle',
+    publishedOn : '3/10/2016',
+    body: '<p>Commuters got it BAD...we really do!</p>'
+  },
+  {
+    title: 'Global Warming',
+    blogcategory: 'World News',
+    publishedOn : '2/10/2016',
+    body: '<p>Our dying planet</p>'
+  },
+  {
+    title: 'Lesser of two Evils',
+    blogcategory: 'Politics',
+    publishedOn : '1/10/2016',
+    body: '<p>VOTE VOTE VOTE</p>'
+  }
+];
