+++ public/scripts/blogs.js
