+++ public/scripts/mainView.js
