+++ public/styles/base.css
@@ -81,7 +81,7 @@ p{
   font-style: normal;
   text-align: left;
   text-indent: 0px;
-  font-size: 22px;
+  font-size: 1.5em;
   padding: 15px;
   margin: 30px;
   width: 90%;
