+++ public/styles/icons.css
