--- scripts/blogArticles.js
@@ -1,59 +1,0 @@
-'use strict';
-var blogArticles = [
-  {
-    title:'Migos',
-    body:'<p>Bad and Boujee</p>\n',
-    projectUrl:'https://www.youtube.com/watch?v=S-sJp1FfG7Q',
-    category:'Music',
-    publishedOn:'2017-01-10T01:13:54.102Z','daysAgo':0,
-    publishedStatus:'published 0 days ago'
-  },
-  {
-    title: 'My Journey',
-    blogcategory: 'Finances',
-    publishedOn : '10/10/2016',
-    body: '<p>Hello World!  My name is Jeremiah Walters I am a proud father and committed student in need of financial support to complete my education as a coding and computer programming student.</p>'
-  },
-  {
-    title: 'J. Cole',
-    blogcategory: 'Music & Entertainment',
-    publishedOn : '7/10/2016',
-    body: '<p>Jermaine Lamarr Cole (born January 28, 1985), known professionally as J. Cole, is an American rapper, songwriter and record producer.</p>'
-  },
-  {
-    title: 'Seattle Start Ups',
-    blogcategory: 'Tech',
-    publishedOn : '6/10/2016',
-    body: '<p>Seattle startup culture is booming and its good for people who are interested in running a business that allows them to be creative with their approach.</p>'
-  },
-  {
-    title: '1st 5k',
-    blogcategory: 'Health & Exercise',
-    publishedOn : '5/10/2016',
-    body: '<p>Running is good for your heart:)</p>'
-  },
-  {
-    title: 'No More Chicken Nuggets',
-    blogcategory: 'Health & Exercise',
-    publishedOn : '4/10/2016',
-    body: '<p>eat better live longer</p>'
-  },
-  {
-    title: 'Traffic Sucks! Can I Just Stay Home?',
-    blogcategory: 'Lifestyle',
-    publishedOn : '3/10/2016',
-    body: '<p>Commuters got it BAD...we really do!</p>'
-  },
-  {
-    title: 'Global Warming',
-    blogcategory: 'World News',
-    publishedOn : '2/10/2016',
-    body: '<p>Our dying planet</p>'
-  },
-  {
-    title: 'Lesser of two Evils',
-    blogcategory: 'Politics',
-    publishedOn : '1/10/2016',
-    body: '<p>VOTE VOTE VOTE</p>'
-  }
-];
