--- scripts/projectData.js
@@ -1,31 +1,0 @@
-// 'use strict';
-// var projectData = [
-//   {'title':'How to Make It Pretty','body':'<pre><code>h1{\nbackground-color: Black;\n}\n</code></pre>','projectUrl':'google.com','category':'Tech','publishedOn':'2017-01-10T01:00:08.180Z','daysAgo':0,'publishedStatus':'published 0 days ago'},
-//   {
-//     title: 'Linked Up',
-//     category:'Networking',
-//     projectUrl:'https://linkedupdep.herokuapp.com',
-//     publishedOn: '2016-12-06',
-//     body: '<p>Meet up with potential business partners, employers or employees using a combination of linkedin and meet up.com apis. Link up based on topic, location and dates. Find people who share a passion with you and you never know you could end up growing your network or landing a great job or business partner from face to face professional interactions.</p>'
-//   },
-//   {
-//     title:  'PNW Taste',
-//     category: 'Food',
-//     projectUrl: 'https://github.com/jtwalters25/PNW-Taste',
-//     publishedOn: '2016-11-05',
-//     body: '<p>Pacific Northwest Taste Restaurant Site: A website for a new restaurant that serves Pacific Northwest cusine and delicious drinks.  Vanilla JS is used in this code with OOP, functions and constructors.  Go make a reservation today on the reservation page or find out the Happy Hour drink special with one click!</p>'
-//   },
-//   {
-//     title: 'Guessing Game',
-//     category: 'Gaming',
-//     projectUrl:'https://jeremiahsguessinggame.herokuapp.com/',
-//     publishedOn:'2016-10-10',
-//     body: '<p>Guess a number between 0 and 100.  If you are right you will be alerted.  You will recieve messages indicating if you are too low or too high</p>'
-//   },
-//   {
-//     title: 'Get There',
-//     category: 'Transportation',
-//     publishedOn: '2016-09-07',
-//     body: '<p>This navigation gets you there with routes that us alleys and backstreets that only a local or Seattle native would know</p>'
-//   },
-// ];
